worship/requirements.txt
attrs==21.4.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
identify==2.5.1
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
nodeenv==1.7.0
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.1
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
virtualenv==20.15.0
Werkzeug==2.1.2
XlsxWriter==3.0.3
zipp==3.8.0
